+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/webview/in_app_webview/InAppWebViewSettings.java
@@ -49,6 +49,8 @@ public class InAppWebViewSettings implements ISettings<InAppWebViewInterface> {
   public List<Map<String, Map<String, Object>>> contentBlockers = new ArrayList<>();
   public Integer preferredContentMode = PreferredContentModeOptionType.RECOMMENDED.toValue();
   public Boolean useShouldInterceptAjaxRequest = false;
+  public Boolean useOnAjaxReadyStateChange = false;
+  public Boolean useOnAjaxProgress = false;
   public Boolean interceptOnlyAsyncAjaxRequests = true;
   public Boolean useShouldInterceptFetchRequest = false;
   public Boolean incognito = false;
@@ -218,6 +220,12 @@ public class InAppWebViewSettings implements ISettings<InAppWebViewInterface> {
           break;
         case "useShouldInterceptAjaxRequest":
           useShouldInterceptAjaxRequest = (Boolean) value;
+          break;
+        case "useOnAjaxReadyStateChange":
+          useOnAjaxReadyStateChange = (Boolean) value;
+          break;
+        case "useOnAjaxProgress":
+          useOnAjaxProgress = (Boolean) value;
           break;
         case "interceptOnlyAsyncAjaxRequests":
           interceptOnlyAsyncAjaxRequests = (Boolean) value;
@@ -497,6 +505,8 @@ public class InAppWebViewSettings implements ISettings<InAppWebViewInterface> {
     settings.put("contentBlockers", contentBlockers);
     settings.put("preferredContentMode", preferredContentMode);
     settings.put("useShouldInterceptAjaxRequest", useShouldInterceptAjaxRequest);
+    settings.put("useOnAjaxReadyStateChange", useOnAjaxReadyStateChange);
+    settings.put("useOnAjaxProgress", useOnAjaxProgress);
     settings.put("interceptOnlyAsyncAjaxRequests", interceptOnlyAsyncAjaxRequests);
     settings.put("useShouldInterceptFetchRequest", useShouldInterceptFetchRequest);
     settings.put("incognito", incognito);