+++ flutter_inappwebview_ios/ios/Classes/InAppWebView/InAppWebViewSettings.swift
@@ -27,6 +27,8 @@ public class InAppWebViewSettings: ISettings<InAppWebView> {
     var contentBlockers: [[String: [String : Any]]] = []
     var minimumFontSize = 0
     var useShouldInterceptAjaxRequest = false
+    var useOnAjaxReadyStateChange = false
+    var useOnAjaxProgress = false
     var interceptOnlyAsyncAjaxRequests = true
     var useShouldInterceptFetchRequest = false
     var incognito = false