+++ flutter_inappwebview_windows/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -1536,13 +1536,13 @@ class WindowsInAppWebViewController extends PlatformInAppWebViewController
 
               if (webviewParams != null &&
                   webviewParams!.onAjaxReadyStateChange != null)
-                return (await webviewParams!.onAjaxReadyStateChange!(
+                return jsonEncode((await webviewParams!.onAjaxReadyStateChange!(
                         _controllerFromPlatform, request))
-                    ?.toNativeValue();
+                    ?.toNativeValue());
               else
-                return (await _inAppBrowserEventHandler!
+                return jsonEncode((await _inAppBrowserEventHandler!
                         .onAjaxReadyStateChange(request))
-                    ?.toNativeValue();
+                    ?.toNativeValue());
             }
             return null;
           case "onAjaxProgress":
@@ -1555,13 +1555,13 @@ class WindowsInAppWebViewController extends PlatformInAppWebViewController
 
               if (webviewParams != null &&
                   webviewParams!.onAjaxProgress != null)
-                return (await webviewParams!.onAjaxProgress!(
+                return jsonEncode((await webviewParams!.onAjaxProgress!(
                         _controllerFromPlatform, request))
-                    ?.toNativeValue();
+                    ?.toNativeValue());
               else
-                return (await _inAppBrowserEventHandler!
+                return jsonEncode((await _inAppBrowserEventHandler!
                         .onAjaxProgress(request))
-                    ?.toNativeValue();
+                    ?.toNativeValue());
             }
             return null;
           case "shouldInterceptFetchRequest":